refactor of database design to remove configuration
after discussion with @becker33 we decided to remove the model
of a configuration, as it is just a list of packages. We can preserve
the same relationships and metadata by just having a package be at the
top level. This update also enforces creation of the spec file dumps
using a full hash instead of a build hash, as this will be needed
to ensure uniqueness of dependencies

Signed-off-by: vsoch <vsoch@users.noreply.github.com>