start of work to add Build table
We need to include BuildEnvironment and objects alongside a Spec build,
and we need it to be possible to build the same spec in different environments
and have different ids/rows in the table. This is start of work to do that,
and also clean up some of the API namespace. I next need to go into spack
and refactor for this structure, and figure out how to get a list of object
files after a build to (eventually) parse and send to the spack monitor

Signed-off-by: vsoch <vsoch@users.noreply.github.com>