adding logparser to build models
This will add BuildWarning and BuildError to be parsed from a buildPhase.output and buildPhase.error,
and the results are generated on demand and stored in the database to parse into the build details
view. Since we could not easily see the dependencies or parents of a particular spec these are also
added to the main page for the spec details.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>