adding better filtering for matrices
currently we are trying to do multiple queries to get metrics for builds across compilers, arches,
and versions. This is not super efficient (the loop is about 8 seconds) for a web interface. By adding
the dimension of arch and then using a pandas data frame to calculate metrics, we get this down to unles
*under a second

Signed-off-by: vsoch <vsoch@users.noreply.github.com>